videothumbs/deploy/requirements.txt
Django==1.6
django-extensions==1.2.5
django-quicky==0.6.3
django-redis-sessions==0.4.0
path.py==5.0
pytest==2.4.2
redis==2.8.0
fabtools==0.16.0
django-redis-cache==0.10.0
